// src/api/models/structuredPortfolioResDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export interface StructuredPortfolioResDTO {
  portfolioId: number;
  name: string;
  description: string;
  responsibilities: string;
  problemSolving: string;
  learnings: string;
}